SSLEngineFactory is a single implementation
